requirements.txt
fastapi
uvicorn[standard]
openai
pinecone
pydantic>=2
python-dotenv